+++ Code/Graphic/GUI/MessageDialogBox.cpp
@@ -6,8 +6,8 @@
 int MessageDialogBox::m_instanceExists = 0;
 
 MessageDialogBox::MessageDialogBox(ska::TimeObservable& timeObs, ska::MouseObservable& gui, ska::KeyObservable& keyboardObs, ska::InputContextManager& icm, const std::string& text, const std::string& imageResource, const unsigned int screenH, const unsigned int screenW, const int timeout) :
-m_dialog(timeObs, gui, keyboardObs, text, { 0, screenH - TAILLEBLOCFENETRE * 4, screenW / 2, TAILLEBLOCFENETRE * 4 }, timeout),
-m_talker(timeObs, gui, keyboardObs, "", imageResource, { TAILLEBLOCFENETRE, screenH - TAILLEBLOCFENETRE * 9, 4 * TAILLEBLOCFENETRE, 5 * TAILLEBLOCFENETRE }, timeout),
+m_dialog(timeObs, gui, keyboardObs, text, { 0, static_cast<int>(screenH - TAILLEBLOCFENETRE * 4), static_cast<int>(screenW / 2), TAILLEBLOCFENETRE * 4 }, timeout),
+m_talker(timeObs, gui, keyboardObs, "", imageResource, { TAILLEBLOCFENETRE, static_cast<int>(screenH - TAILLEBLOCFENETRE * 9), 4 * TAILLEBLOCFENETRE, 5 * TAILLEBLOCFENETRE }, timeout),
 m_playerICM(icm) {
 	m_instanceExists++;
 }
